@@ -53,7 +53,7 @@ namespace nc::edac
         /// @param n integer value
         /// @return bool true if value is a power of two, else false
         ///
-        template<typename IntType, std::enable_if_t<std::is_integral_v<IntType>, int> = 0>
+        template<std::integral IntType>
         constexpr bool isPowerOfTwo(IntType n) noexcept
         {
             // Returns true if the given non-negative integer n is a power of two.
@@ -72,7 +72,7 @@ namespace nc::edac
         /// @return next power of two
         /// @exception std::invalid_argument if input value is less than zero
         ////
-        template<typename IntType, std::enable_if_t<std::is_integral_v<IntType>, int> = 0>
+        template<std::integral IntType>
         std::size_t nextPowerOfTwo(IntType n)
         {
             if (n < 0)
@@ -96,7 +96,7 @@ namespace nc::edac
         /// @return first n powers of two
         /// @exception std::bad_alloc if unable to allocate for return vector
         ///
-        template<typename IntType, std::enable_if_t<std::is_integral_v<IntType>, int> = 0>
+        template<std::integral IntType>
         std::vector<std::size_t> powersOfTwo(IntType n)
         {
             auto i      = std::size_t{ 0 };
@@ -123,7 +123,7 @@ namespace nc::edac
         /// @exception std::invalid_argument if input value is less than zero
         /// @exception std::runtime_error if the number of data bits does not represent a valid Hamming SECDED code
         ///
-        template<typename IntType, std::enable_if_t<std::is_integral_v<IntType>, int> = 0>
+        template<std::integral IntType>
         std::size_t numSecdedParityBitsNeeded(IntType numDataBits)
         {
             const auto n               = nextPowerOfTwo(numDataBits);
@@ -152,10 +152,7 @@ namespace nc::edac
         /// @exception std::invalid_argument if parityBit is not a power of two
         /// @exception std::bad_alloc if unable to allocate return vector
         ///
-        template<typename IntType1,
-                 typename IntType2,
-                 std::enable_if_t<std::is_integral_v<IntType1>, int> = 0,
-                 std::enable_if_t<std::is_integral_v<IntType2>, int> = 0>
+        template<std::integral IntType1, std::integral IntType2>
         std::vector<std::size_t> dataBitsCovered(IntType1 numDataBits, IntType2 parityBit)
         {
             if (!isPowerOfTwo(parityBit))
@@ -233,7 +230,7 @@ namespace nc::edac
         /// @exception std::invalid_argument if parityBit is not a power of two
         /// @exception std::bad_alloc if unable to allocate return vector
         ///
-        template<std::size_t DataBits, typename IntType, std::enable_if_t<std::is_integral_v<IntType>, int> = 0>
+        template<std::size_t DataBits, std::integral IntType>
         bool calculateParity(const std::bitset<DataBits>& data, IntType parityBit)
         {
             const auto bitsCovered = dataBitsCovered(DataBits, parityBit);
@@ -251,9 +248,8 @@ namespace nc::edac
         /// @exception std::runtime_error if DataBits and EncodedBits are not consistent
         /// @exception std::runtime_error if the number of data bits does not represent a valid Hamming SECDED code
         ///
-        template<std::size_t DataBits,
-                 std::size_t EncodedBits,
-                 std::enable_if_t<greaterThan_v<EncodedBits, DataBits>, int> = 0>
+        template<std::size_t DataBits, std::size_t EncodedBits>
+        requires GreaterThan<EncodedBits, DataBits>
         std::size_t checkBitsConsistent()
         {
             const auto numParityBits = detail::numSecdedParityBitsNeeded(DataBits);
@@ -273,9 +269,8 @@ namespace nc::edac
         /// @param encodedBits the Hamming SECDED encoded word
         /// @return data bits from the encoded word
         ///
-        template<std::size_t DataBits,
-                 std::size_t EncodedBits,
-                 std::enable_if_t<greaterThan_v<EncodedBits, DataBits>, int> = 0>
+        template<std::size_t DataBits, std::size_t EncodedBits>
+        requires GreaterThan<EncodedBits, DataBits>
         std::bitset<DataBits> extractData(const std::bitset<EncodedBits>& encodedBits) noexcept
         {
             auto dataBits = std::bitset<DataBits>();
@@ -349,9 +344,8 @@ namespace nc::edac
     /// @exception std::runtime_error if DataBits and EncodedBits are not consistent
     /// @exception std::runtime_error if the number of data bits does not represent a valid Hamming SECDED code
     ///
-    template<std::size_t DataBits,
-             std::size_t EncodedBits,
-             std::enable_if_t<greaterThan_v<EncodedBits, DataBits>, int> = 0>
+    template<std::size_t DataBits, std::size_t EncodedBits>
+    requires GreaterThan<EncodedBits, DataBits>
     int decode(std::bitset<EncodedBits> encodedBits, std::bitset<DataBits>& decodedBits)
     {
         const auto numParityBits = detail::checkBitsConsistent<DataBits, EncodedBits>();
