@@ -45,7 +45,7 @@ namespace nc::utils
     ///
     /// @return bool
     ///
-    template<typename dtype, std::enable_if_t<std::is_integral<dtype>::value, int> = 0>
+    template<std::integral dtype>
     bool essentiallyEqual(dtype inValue1, dtype inValue2) noexcept
     {
         return inValue1 == inValue2;
@@ -61,7 +61,7 @@ namespace nc::utils
     ///
     /// @return bool
     ///
-    template<typename dtype, std::enable_if_t<std::is_floating_point<dtype>::value, int> = 0>
+    template<std::floating_point dtype>
     bool essentiallyEqual(dtype inValue1, dtype inValue2, dtype inEpsilon) noexcept
     {
         const auto absValue1 = std::abs(inValue1);
@@ -78,7 +78,7 @@ namespace nc::utils
     ///
     /// @return bool
     ///
-    template<typename dtype, std::enable_if_t<std::is_floating_point<dtype>::value, int> = 0>
+    template<std::floating_point dtype>
     bool essentiallyEqual(dtype inValue1, dtype inValue2) noexcept
     {
         return essentiallyEqual(inValue1, inValue2, DtypeInfo<dtype>::epsilon());
